requirements.txt
langchain
langgraph
generative-ai-hub-sdk
python-dotenv==1.0.1
fastmcp
mcp
requests
fastapi
uvicorn
httpx
langchain-mcp-adapters
langchain-aws
langchain-community
aioboto3
google-api-core
google-cloud-aiplatform
langchain-openai
langchain-text-splitters
langchain_aws
langchain-google-vertexai
langchain_mcp_adapters
sqlalchemy>=2.0.0
alembic>=1.14.0
python-dateutil>=2.8.0
